refactor(mobileapp): rename Mission files to Job naming
