+++ mobileapp/src/components/organisms/JobsList.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { FlatList } from "react-native";
 import { FAKE_MISSIONS_DATA } from "../../common/mockedData";
-import MissionCard from "../molecules/MissionCard";
+import MissionCard from "../molecules/JobCard";
 
-export default function MissionsList() {
+export default function JobsList() {
   // TODO: Add this hook.
   // const { missions } = useMissions();
