+++ mobileapp/src/navigation/StackNavigationParams.ts
@@ -1,7 +1,7 @@
 import type { ProfileType } from "../screens/single/Profile";
 
 export type StackParamList = {
-  Mission: { id: number };
+  Job: { id: number };
   Conversation: { id: number };
   Profile: { id: number; type: ProfileType };
 } & AuthParamList &
@@ -16,13 +16,13 @@ export type AuthParamList = {
 };
 
 export type AppParamList = {
-  // TODO: Add filters param for Missions page.
-  Missions: undefined;
+  // TODO: Add filters param for Jobs page.
+  Jobs: undefined;
   Boutique: undefined;
   Profil: undefined;
   Messagerie: undefined;
   Scanner: undefined;
-  MissionsFilter: undefined;
+  JobsFilter: undefined;
 };
 
 export type BottomTabParamList = {
